Modularize redaction script
