+++ freedact/__main__.py
@@ -1,0 +1,8 @@
+"""Module executed when running ``python -m freedact``."""
+
+from .cli import main
+
+
+if __name__ == "__main__":
+    main()
+
